[NRT-591] Extract BulkSuggestionSummaryDialog into its own module
- Move the bulk summary dialog, its action-state machine, categories, color
  helpers, and the bulk-submit callback to gui/bulk_suggestion_summary_dialog.py.
- Move the shared panel color helpers to gui/utils.py (avoids a circular
  import between the two dialog modules).
- Drop the change_type plumbing from the bulk callback/dialog - converted
  resubmits are always 'Updated content', matching the single-note path.
- Flatten the nested resubmit task/on_done callbacks in
  _show_note_already_exists_dialog.
- Rename module constants to the no-underscore convention.
